+++ src/features/commands.ts
@@ -3,3 +3,4 @@
 // Module to hold all command names
 
 export const RestartLS = 'fortran.analysis.restartLanguageServer';
+export const RescanLint = 'fortran.analysis.rescanLinter';